Checkout transaction is add

+++ src/main/java/com/Integration/NTI/Controllers/BookController.java
@@ -1,9 +1,12 @@
 package com.Integration.NTI.Controllers;
 
 import com.Integration.NTI.Models.Book;
+import com.Integration.NTI.Models.Role;
+import com.Integration.NTI.Models.User;
 import com.Integration.NTI.Repositries.UserRepo;
 import com.Integration.NTI.Requests.CartRequest;
 import com.Integration.NTI.Services.BookService;
+import com.Integration.NTI.Services.UserService;
 import com.paypal.api.payments.Payment;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -23,21 +26,22 @@ public class BookController {
     private BookService bookService;
     private UserRepo userRepo;
 
+    private UserService userService;
+
 
     @Autowired
-    public BookController(BookService bookService, UserRepo userRepo) {
+    public BookController(BookService bookService, UserRepo userRepo, UserService userService) {
         this.bookService = bookService;
         this.userRepo = userRepo;
+        this.userService = userService;
     }
    @PostMapping("/create")
    @PreAuthorize("hasRole('ADMIN')")
    public ResponseEntity<String> createNewBook(@RequestBody Book book){
-       Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-       boolean isAdmin = authentication.getAuthorities().stream()
-               .anyMatch(auth -> auth.getAuthority().equals("ROLE_ADMIN"));
 
+       User user = userService.returnUserAuth();
        // Print the user's role in the console
-       if (isAdmin) {
+       if (user.getRoles().contains(Role.ADMIN)) {
            bookService.addBook(book);
            return new ResponseEntity<>("BOOK IS ADDED SUCCESSFULLY...", HttpStatus.CREATED);
 
@@ -45,7 +49,6 @@ public class BookController {
            return new ResponseEntity<>("ONLY ADMINS CAN ADD NEW BOOK...", HttpStatus.UNAUTHORIZED);
        }
    }
-
     @GetMapping({"/all", "/"})
     public ResponseEntity<List<Book>> listBooks(){
         List<Book> list =  bookService.findAll();
@@ -59,7 +62,15 @@ public class BookController {
     @GetMapping("/delete/{id}")
 
     public ResponseEntity<String> deleteById(@RequestBody CartRequest request){
+        User user = userService.returnUserAuth();
+        // Print the user's role in the console
+        if (user.getRoles().contains(Role.ADMIN)) {
         bookService.deleteById(request.getBookId(), request.getQuantity());
         return new ResponseEntity<>("BOOK IS SUCCESSFULLY DELETED",HttpStatus.NO_CONTENT);
+        }
+        else{
+            return new ResponseEntity<>("ONLY ADMINS CAN DELETE BOOK...", HttpStatus.UNAUTHORIZED);
+
+        }
     }
 }
